requirements.txt
numpy<1.26.5,>=1.26
numpy-quaternion>=2023.0.2
pyquaternion<1.0.0
python-osc<1.9.0
bleak<0.21,>=0.15.1
